Create game.py
This program allows you to practice aiming for FPS games. At the beginning there is a countdown from 3 to 0, and then the game board appears. In the upper left corner we have a click counter, and in the right corner the time it took us to click the dot, and below that the average response time